requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
python-dotenv==1.0.0
boto3==1.35.10
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
celery==5.4.0
